[Code] Cleanup Strings Header
